Update init.lua
Fixed the crashing for both pday and slowmo, and removed the blind command. The only real issue with it is that I wasn't able to get the blind/tanning status effects to stick when a player quits, or to show in the /curses status list, so I removed the relevant code. Butttt It works. When I learn more, I'll probably return to the issue.